+++ vaultron.tf
@@ -9,12 +9,12 @@
 
 # Set TF_VAR_vault_version to set this
 variable "vault_version" {
-  default = "0.11.4"
+  default = "0.11.5"
 }
 
 # Set TF_VAR_consul_version to set this
 variable "consul_version" {
-  default = "1.3.0"
+  default = "1.4.0"
 }
 
 # -----------------------------------------------------------------------